+++ java/src/main/java/com/techelevator/controller/PostController.java
@@ -1,11 +1,17 @@
 package com.techelevator.controller;
 
+import com.techelevator.dao.ModerationDao;
 import com.techelevator.dao.PostDao;
+import com.techelevator.dao.UserDao;
+import com.techelevator.model.Moderation;
+import com.techelevator.model.User;
 import com.techelevator.model.request.CreatePostDto;
 import com.techelevator.model.responses.PostResponseDto;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.security.Principal;
 import java.util.List;
@@ -17,6 +23,10 @@ public class PostController {
 
     @Autowired
     PostDao postDao;
+    @Autowired
+    UserDao userDao;
+    @Autowired
+    ModerationDao moderationDao;
 
     // public List<PostResponseDto> getPosts(long forum, String keyword, int limit, boolean sorBytPopularity)
     @GetMapping("/posts")
@@ -58,10 +68,16 @@ public class PostController {
     }
 
 
-    @PreAuthorize("hasAnyRole('ADMIN', 'MODERATOR')") //ToDo from: jasir to self change logic for this
+    @PreAuthorize("isAuthenticated()")
     @DeleteMapping("posts/{id}/delete")
-    public void deletePost(@PathVariable long id, Principal user){
-        postDao.deletePost(id);
+    public void deletePost(@PathVariable long id, Principal user) {
+        if (checkUserRole(id, user.getName())) {
+            postDao.deletePost(id);
+        }
+        else {
+            //throw an exception 401 or forbidden
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
     }
 
     @PreAuthorize("isAuthenticated()")
@@ -83,5 +99,35 @@ public class PostController {
     @PutMapping("posts/{id}/downvote/undislike")
     public void unvotingDislike(@PathVariable long id) {
         postDao.unvote(id, 1);
+    }
+
+
+    public boolean checkUserRole(long postId, String username) {
+        boolean hasPermission = false;
+        boolean isAMod = false;
+        boolean isAAdmin = false;
+
+        PostResponseDto post = getPostById(postId);
+        List<Moderation> moderator = moderationDao.getListOfModeratorsOfForum(post.getForum_id());
+        User user = userDao.getUserByUsername(username);
+
+        if (user.getAuthorities().contains("ROLE_ADMIN")) {
+            isAAdmin = true;
+        }
+
+        for (Moderation mod : moderator) {
+            if (mod.getUsername().equals(username)) {
+                isAMod = true;
+                break;
+            }
+        }
+        if (post.getCreator_username().equals(username)){
+            hasPermission = true;
+        } else if (isAMod) {
+            hasPermission = true;
+        } else if (isAAdmin) {
+            hasPermission = true;
+        }
+        return hasPermission;
     }
 }
